[MOVE] BranchedExecutorを別ファイルに移動する
